rename: store 위치 변경
